tests/wpt/web-platform-tests/css/css-animations/parsing/animation-play-state-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: parsing animation-play-state with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-animations/#propdef-animation-play-state">
<meta name="assert" content="animation-play-state supports only the grammar '<single-animation-play-state> #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("animation-play-state", "auto");
test_invalid_value("animation-play-state", "paused running");
</script>
</body>
</html>
